requirements_utf8.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==4.2.28
django-allauth==65.14.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0

